+++ src/main/java/tn/esprit/pidev/services/GestionMessageImpl.java
@@ -25,7 +25,7 @@ public class GestionMessageImpl implements IGestionMessage {
     @Autowired
     IDiscussionRepository iDiscussionRepository;
     @Override
-    public Message sendMessage(Long userSender, Long discussion, String message) {
+    public Message sendMessage(int userSender, Long discussion, String message) {
         Message messageo = new Message();
         messageo.setMessage(message);
 
@@ -61,7 +61,7 @@ public class GestionMessageImpl implements IGestionMessage {
     }
 
     @Override
-    public Message replyMessage(Long userSender, Long discussion,Long message, String reply) {
+    public Message replyMessage(int userSender, Long discussion,Long message, String reply) {
         Message messageo = new Message();
         messageo.setMessage(reply);
         messageo.setDateSent(LocalDateTime.now());
