+++ src/main/java/tn/esprit/pidev/services/IGestionDiscussion.java
@@ -11,23 +11,23 @@ import tn.esprit.pidev.repository.IUserRepository;
 import java.util.List;
 
 public interface IGestionDiscussion {
-    public Discussion startDiscussionDuo(Long userStart, Long userEnd);
+    public Discussion startDiscussionDuo(int userStart, int userEnd);
 
-    public Discussion startDiscussionGroup(Long userStart, String title, String userList, String image);
+    public Discussion startDiscussionGroup(int userStart, String title, String userList, String image);
 
-    public Discussion startDiscussionCommunity(Long userStart, String title, String userList, String discussionList, String image);
+    public Discussion startDiscussionCommunity(int userStart, String title, String userList, String discussionList, String image);
 
-    public ResponseEntity<String> modifyDiscussionGroup(Long discussion, Long userStart, String title, String userList, Long admin, String image);
+    public ResponseEntity<String> modifyDiscussionGroup(Long discussion, int userStart, String title, String userList, int admin, String image);
 
-    public ResponseEntity<String> modifyDiscussionCommunity(Long discussion, Long userStart, String title, String userList, String discussionList, Long admin, String image);
+    public ResponseEntity<String> modifyDiscussionCommunity(Long discussion, int userStart, String title, String userList, String discussionList, int admin, String image);
 
-    public ResponseEntity<String> addAdminsToDiscussion(Long discussion, Long admin, String userList);
+    public ResponseEntity<String> addAdminsToDiscussion(Long discussion, int admin, String userList);
 
-    public List<Discussion> retrieveAllDiscussions(Long id);
+    public List<Discussion> retrieveAllDiscussions(Integer id);
 
-    public List<Discussion> retrieveAllCommunities(Long id);
+    public List<Discussion> retrieveAllCommunities(int id);
 
-    public Discussion leaveDiscussion(Long user, Long discussion);
+    public Discussion leaveDiscussion(int user, Long discussion);
 
-    public ResponseEntity<String> deleteDiscussion(Long user, Long discussion);
+    public ResponseEntity<String> deleteDiscussion(int user, Long discussion);
 }
