+++ src/main/java/tn/esprit/pidev/services/IGestionReaction.java
@@ -4,6 +4,6 @@ import tn.esprit.pidev.entities.Reaction;
 
 public interface IGestionReaction {
 
-    public Reaction reacting (Long user, Long message, String reaction);
+    public Reaction reacting (int user, Long message, String reaction);
 
 }
